Adding the R1 directory for the manuscript revision and including reviewed scripts
Scripts are being modified to adopt more recent and faster solutions.
This commit deals with the spatial dataset preparation.
The main changes to consisted in splitting functions and code across various files, in implementing simple features for spatial manipulation, and in removing the environmental dataset from the spatial manipulation script.